chore: move FVT-only variables to fvt files
Move this into their own go:build functional files so they do not appear
"unused" to golangci-lint
